Move to utils, add doc comments
@12944qwerty
