+++ src/renderer/util.ts
@@ -11,3 +11,61 @@ export const loadStyleSheet = (path: string): HTMLLinkElement => {
 
   return el;
 };
+
+export type ReactFiber<T extends Record<string, unknown> = Record<string, unknown>> = T & {
+  stateNode: Element | T;
+  return: ReactFiber<T>;
+};
+
+/**
+ * Wait for an element to be added to the DOM
+ * @param selector Element selector
+ */
+export async function waitFor(selector: string): Promise<Element> {
+  let element: Element | null = null;
+
+  while (!element) {
+    element = document.querySelector(selector);
+    await new Promise((resolve) => requestAnimationFrame(resolve));
+  }
+
+  return element;
+}
+
+/**
+ * Get the React instance of an element
+ * @param element Element to get the React instance of
+ * @returns React instance
+ * @throws If the React instance could not be found
+ */
+export function getReactInstance<T extends Record<string, unknown> = Record<string, unknown>>(
+  element: Element,
+): ReactFiber<T> {
+  const keys = Object.keys(element);
+  const reactKey = keys.find((key) => key.startsWith("__reactFiber$"));
+  if (!reactKey) {
+    throw new Error("Could not find react fiber");
+  }
+  // @ts-expect-error Doesn't like the dynamic key I guess
+  return element[reactKey];
+}
+
+/**
+ * Get the React owner instance of an element
+ * @param element Element to get the React owner instance of
+ * @returns React owner instance
+ * @throws If the React owner instance could not be found
+ */
+export function getOwnerInstance<T extends Record<string, unknown> = Record<string, unknown>>(
+  element: Element,
+): T {
+  let current = getReactInstance<T>(element);
+  while (current) {
+    const owner = current.stateNode;
+    if (owner && !(owner instanceof Element)) {
+      return owner;
+    }
+    current = current.return;
+  }
+  throw new Error("Could not find react owner");
+}
